# docs/openapi/tenant-openapi.yaml
openapi: 3.1.0
info:
  title: CloudAPI Tenant
  description: '

    Welcome to CloudAPI Tenant!


    For detailed guidance on using the API, please visit our official documentation:

    https://www.didx.co.za/acapy-cloud/index.html

'
  version: 5.0.0-rc5
servers:
  - url: /tenant
paths:
...
  /v1/issuer/credentials:
    post:
      tags:
        - issuer
      summary: Send Holder a Credential
      description: "Create and send a credential, automating the issuer-side flow\n---\nNB: Only a tenant with the issuer role can send credentials.\n\nWhen sending a credential, the credential type must be one of `anoncreds` or `ld_proof`.\n```json\n{\n    \"anoncreds_credential_detail\": {...},\n    \"ld_credential_detail\": {...},\n    \"save_exchange_record\": true, <-- Whether the credential exchange record should be saved on completion.\n    \"connection_id\": \"string\", <-- The issuer's reference to the connection they want to submit the credential to.\n}\n```\nSetting the `save_exchange_record` field to True will save the exchange record after the credential is accepted.\nThe default behaviour is to only save exchange records while they are in pending state.\n\nFor a detailed technical specification of the credential issuing process, refer to the [Aries Issue Credential v2\nRFC](https://github.com/hyperledger/aries-rfcs/blob/main/features/0453-issue-credential-v2/README.md).\n\nRequest Body:\n---\n    credential: SendCredential\n        The payload for sending a credential\n\nReturns\n-------\n    CredentialExchange\n        A record of this credential exchange"
      operationId: send_credential_v1_issuer_credentials_post
      security:
        - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCredential'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialExchange'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - issuer
      summary: Fetch Credential Exchange Records
      description: "Get a list of credential exchange records\n---\nBoth holders and issuers can call this endpoint, because they each have their own records of a credential exchange.\n\nThese records contain information about the credentials issued to a holder, such as the current state of the\nexchange, and other metadata such as the `connection_id` that a credential was submit to (if an issuer) or received\nfrom (if a holder). Each record in the list is related to a single credential exchange flow.\n\nNB: An issuer and a holder will have distinct credential exchange ids, despite referring to the same exchange.\nThe `thread_id` is the only record attribute that will be the same for the holder and the issuer.\n\nAn exchange record will automatically be deleted after a flow completes (i.e. when state is 'done'),\nunless the `save_exchange_record` was set to true.\n\nThe following parameters can be set to filter the fetched exchange records: connection_id, role, state, thread_id.\n\nParameters (Optional):\n---\n    limit: int - The maximum number of records to retrieve\n    offset: int - The offset to start retrieving records from\n    descending: bool - Whether to return results in descending order. Results are ordered by record created time.\n    connection_id: str\n    role: Role: \"issuer\", \"holder\"\n    state: State: \"proposal-sent\", \"proposal-received\", \"offer-sent\", \"offer-received\",\n                             \"request-sent\", \"request-received\", \"credential-issued\", \"credential-received\",\n                             \"credential-revoked\", \"abandoned\", \"done\"\n    thread_id: UUID\n\nReturns\n-------\n    List[CredentialExchange]\n        A list of credential exchange records"
      operationId: get_credentials_v1_issuer_credentials_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                maximum: 10000
                minimum: 1
              - type: 'null'
            description: Number of results to return
            default: 1000
            title: Limit
          description: Number of results to return
        - name: offset
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
                minimum: 0
              - type: 'null'
            description: Offset for pagination
            default: 0
            title: Offset
          description: Offset for pagination
        - name: descending
          in: query
          required: false
          schema:
            type: boolean
            description: Whether to return results in descending order. Results are ordered by record created time.
            default: true
            title: Descending
          description: Whether to return results in descending order. Results are ordered by record created time.
        - name: connection_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Connection Id
        - name: role
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - issuer
                  - holder
                type: string
              - type: 'null'
            title: Role
        - name: state
          in: query
          required: false
          schema:
            anyOf:
              - enum:
                  - proposal-sent
                  - proposal-received
                  - offer-sent
                  - offer-received
                  - request-sent
                  - request-received
                  - credential-issued
                  - credential-received
                  - credential-revoked
                  - abandoned
                  - done
                  - deleted
                type: string
              - type: 'null'
            title: State
        - name: thread_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialExchange'
                title: Response Get Credentials V1 Issuer Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/issuer/credentials/revoke:
    post:
      tags:
        - revocation
      summary: Revoke a Credential (if revocable)
      description: "Revoke a credential\n---\nRevoke a credential by providing the identifier of the exchange.\n\nIf an issuer is going to revoke more than one credential, it is recommended to set the\n'auto_publish_on_ledger' field to False (default), and then batch publish the revocations using\nthe 'publish-revocations' endpoint.\n\nBy batching the revocations, the issuer can save on transaction fees related to\npublishing revocations to the ledger.\n\nRequest Body:\n---\n    body: RevokeCredential\n        - credential_exchange_id (str): The ID associated with the credential exchange that should be revoked.\n        - auto_publish_on_ledger (bool): (True) publish revocation to ledger immediately, or\n            (default, False) mark it pending\n\nReturns\n-------\n    RevokedResponse:\n        revoked_cred_rev_ids:\n          The revocation registry indexes that were revoked.\n          Will be empty if the revocation was marked as pending."
      operationId: revoke_credential_v1_issuer_credentials_revoke_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeCredential'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /v1/issuer/credentials/revocation/record:
    get:
      tags:
        - revocation
      summary: Fetch a Revocation Record
      description: "Get a credential revocation record\n---\nFetch a credential revocation record by providing the credential exchange id.\nRecords can also be fetched by providing the credential revocation id and revocation registry id.\n\nThe record is the payload of the webhook event on topic 'issuer_cred_rev', and contains the credential's revocation\nstatus and other metadata.\n\nThe revocation registry id (rev_reg_id) and credential revocation id (cred_rev_id) can be found\nin this record if you have the credential exchange id.\n\nParameters\n----------\n    credential_exchange_id: str\n    credential_revocation_id: str\n    revocation_registry_id: str\n\nReturns\n-------\n    IssuerCredRevRecord\n        The credential revocation record\n\nRaises\n------\n    CloudApiException: 400\n        If credential_exchange_id is not provided, both credential_revocation_id and revocation_registry_id must be."
      operationId: get_credential_revocation_record_v1_issuer_credentials_revocation_record_get
      security:
        - APIKeyHeader: []
      parameters:
        - name: credential_exchange_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Credential Exchange Id
        - name: credential_revocation_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Credential Revocation Id
        - name: revocation_registry_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Revocation Registry Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuerCredRevRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/issuer/credentials/publish-revocations:
    post:
      tags:
        - revocation
      summary: Publish Pending Revocations
      description: "Write pending revocations to the ledger\n---\nRevocations that are in a pending state can be published to the ledger.\n\nThe endpoint accepts a `revocation_registry_credential_map`, which provides a dictionary of\nrevocation registry IDs to credential revocation IDs, to allow publishing individual credentials.\n\nIf no revocation registry id is provided (i.e. an empty map `revocation_registry_credential_map: {}`),\nthen all pending revocations will be published.\n\nIf no credential revocation id is provided under a given revocation registry id, then all pending revocations for\nthe given revocation registry id will be published.\n\nWhere to find the revocation registry id and credential revocation id:\nWhen issuing a credential, against a credential definition that supports revocation,\nthe issuer will receive a webhook event on the topic 'issuer_cred_rev'. This event will contain\nthe credential exchange id (cred_ex_id), the credential revocation id (cred_rev_id) and\nthe revocation registry id (rev_reg_id).\n\nRequest Body:\n---\n    publish_request: PublishRevocationsRequest\n        An instance of `PublishRevocationsRequest` containing a `revocation_registry_credential_map`. This map\n        is a dictionary where each key is a revocation registry ID and its value is a list of credential\n        revocation IDs to be published. Providing an empty list for a registry ID instructs the system to\n        publish all pending revocations for that ID. An empty dictionary signifies that all pending\n        revocations across all registry IDs should be published.\n\nReturns\n-------\n    RevokedResponse:\n        revoked_cred_rev_ids:\n          The revocation registry indexes that were revoked.\n          Will be empty if there were no revocations to publish."
      operationId: publish_revocations_v1_issuer_credentials_publish_revocations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRevocationsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /v1/issuer/credentials/clear-pending-revocations:
    post:
      tags:
        - revocation
      summary: Clear Pending Revocations
      description: "Clear pending revocations\n---\n\nNote: This endpoint is not supported for the 'askar-anoncreds' wallet type.\n\nRevocations that are in a pending state can be cleared, such that they are no longer set to be revoked.\n\nThe endpoint accepts a `revocation_registry_credential_map`, which provides a dictionary of\nrevocation registry IDs to credential revocation IDs, to allow clearing individual credentials.\n\nIf no revocation registry id is provided (i.e. an empty map `revocation_registry_credential_map: {}`),\nthen all pending revocations will be cleared.\n\nIf no credential revocation id is provided under a given revocation registry id, then all pending revocations for\nthe given revocation registry id will be cleared.\n\nWhere to find the revocation registry id and credential revocation id:\nWhen issuing a credential, against a credential definition that supports revocation,\nthe issuer will receive a webhook event on the topic 'issuer_cred_rev'. This event will contain\nthe credential exchange id (cred_ex_id), the credential revocation id (cred_rev_id) and\nthe revocation registry id (rev_reg_id).\n\nRequest Body:\n---\n    clear_pending_request: ClearPendingRevocationsRequest\n        An instance of `ClearPendingRevocationsRequest` containing a `revocation_registry_credential_map`. This map\n        is a dictionary where each key is a revocation registry ID and its value is a list of credential\n        revocation IDs to be cleared. Providing an empty list for a registry ID instructs the system to\n        clear all pending revocations for that ID. An empty dictionary signifies that all pending\n        revocations across all registry IDs should be cleared.\n\nReturns\n-------\n    ClearPendingRevocationsResult\n        The revocations that are still pending after the clear request is performed"
      operationId: clear_pending_revocations_v1_issuer_credentials_clear_pending_revocations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearPendingRevocationsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearPendingRevocationsResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /v1/issuer/credentials/get-pending-revocations/{revocation_registry_id}:
    get:
      tags:
        - revocation
      summary: Get Pending Revocations
      description: "Get pending revocations\n---\nGet the pending revocations for a given revocation registry ID.\n\nParameters\n----------\n    revocation_registry_id: str\n        The ID of the revocation registry for which to fetch pending revocations\n\nReturns\n-------\n    PendingRevocations:\n        A list of cred_rev_ids pending revocation for a given revocation registry ID"
      operationId: get_pending_revocations_v1_issuer_credentials_get_pending_revocations__revocation_registry_id__get
      security:
        - APIKeyHeader: []
      parameters:
        - name: revocation_registry_id
          in: path
          required: true
          schema:
            type: string
            title: Revocation Registry Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingRevocations'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PublishRevocationsRequest:
      properties:
        revocation_registry_credential_map:
          additionalProperties:
            items:
              type: string
            type: array
          type: object
          title: Revocation Registry Credential Map
          description: A map of revocation registry IDs to lists of credential revocation IDs that should be published. Providing an empty list for a registry ID publishes all pending revocations for that ID. An empty dictionary signifies that the action should be applied to all pending revocations across all registry IDs.
          default: {}
      type: object
      title: PublishRevocationsRequest
...
    RevokeCredential:
      properties:
        credential_exchange_id:
          type: string
          title: Credential Exchange Id
        auto_publish_on_ledger:
          type: boolean
          title: Auto Publish On Ledger
          default: false
      type: object
      required:
        - credential_exchange_id
      title: RevokeCredential
    RevokedResponse:
      properties:
        cred_rev_ids_published:
          additionalProperties:
            items:
              type: integer
            type: array
          type: object
          title: Cred Rev Ids Published
          description: A map of revocation registry IDs to lists of credential revocation IDs (as integers) that have been revoked.When cred_rev_ids_published is empty no revocations were published.This will happen when revoke is called with auto_publish_on_ledger=False.
      type: object
      title: RevokedResponse
...
    SendCredential:
      properties:
        save_exchange_record:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Save Exchange Record
          description: Controls exchange record retention after exchange is complete. None uses wallet default (typically to delete), true forces save, false forces delete.
        ld_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/LDProofVCDetail'
            - type: 'null'
        anoncreds_credential_detail:
          anyOf:
            - $ref: '#/components/schemas/AnonCredsCredential'
            - type: 'null'
        connection_id:
          type: string
          title: Connection Id
      type: object
      required:
        - connection_id
      title: SendCredential
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-api-key